reorganise sharadar cleaning logic

+++ dbt/sharadar/models/typed/actions_typed.sql
@@ -3,7 +3,7 @@ CAST(strptime(date, '%Y-%m-%d') AS DATE) as date,
 action, 
 ticker,
 name,
-CAST(value as NUMERIC) AS value,
+CAST(value as DOUBLE) AS value,
 contraticker,
 contraname
 FROM
